% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_R.R
\name{get_R}
\alias{get_R}
\alias{get_R.default}
\alias{get_R.integer}
\alias{get_R.numeric}
\alias{get_R.incidence}
\title{Estimate the Reproduction Number}
\usage{
get_R(x, ...)

\method{get_R}{default}(x, ...)

\method{get_R}{integer}(x, disease = "ebola", si_mean = NULL,
  si_sd = NULL, max_R = 5, days = 30)

\method{get_R}{numeric}(x, ...)

\method{get_R}{incidence}(x, ...)
}
\arguments{
\item{x}{An \code{incidence} object, as generated by the package
\code{incidence}, or an \code{integer} vector containing daily incidence.}

\item{disease}{A character string indicating the name of the disease
studied. If provided, then \code{si_mean} and \code{si_sd} will be filled
in automatically using value from the literature.}

\item{si_mean}{The mean of the serial interval distribution.}

\item{si_sd}{The standard deviation of the serial interval distribution.}

\item{max_R}{The maximum value the reproduction number can take.}
}
\description{
This function estimates the (most of the time, 'basic') reproduction number
(R) using i) the known distribution of the Serial Interval (delay between
primary to secondary onset) and ii) incidence data.
}
\examples{

if (require(incidence)) {
 x <- incidence(c(1, 5, 5, 12, 45, 65))
 plot(x)
 res <- get_R(x)
 res
 plot(res)
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
