% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.earlyR.R
\name{plot.earlyR}
\alias{plot.earlyR}
\alias{points.earlyR}
\title{Plot earlyR objects}
\usage{
\method{plot}{earlyR}(x, y = c("R", "lambdas"), scale = 1, ...)

\method{points}{earlyR}(x, scale = 1, ...)
}
\arguments{
\item{x}{A \code{earlyR} object.}

\item{scale}{A numeric value indicating the scaling factor for lambdas.}

\item{...}{Further arguments to be passed to other methods (not used).}
}
\description{
This method is designed for plotting \code{earlyR} objects, output by the
function \code{\link{get_R}}.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
