% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.earlyR.R
\name{print.earlyR}
\alias{print.earlyR}
\title{Print method for earlyR objects}
\usage{
\method{print}{earlyR}(x, ...)
}
\arguments{
\item{x}{A \code{earlyR} object.}

\item{...}{further parameters to be passed to other methods (currently not
used)}
}
\description{
This method prints the content of \code{earlyR} objects.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
